requirements.txt
torch
torchvision
numpy
pandas
matplotlib
Pillow
opencv-python
tqdm